Feature/rest/twitter users stats

+++ code/backend/twitter/examples/postgres/tables.sql
@@ -11,7 +11,7 @@ insert into tweets values(831606548300517376, 6253282, 100, 2);
 
 
 CREATE TABLE "users"(
-    "user_id" integer NOT NULL PRIMARY KEY, 
+    "user_id" numeric NOT NULL PRIMARY KEY, 
     "timestamp" timestamp with time zone NOT NULL DEFAULT CURRENT_TIMESTAMP, 
     "followers" integer NOT NULL, 
     "following" integer NOT NULL
